Only locally override resolvable flag.
Having this in front of the Parallel() call caused this to affect other tests even if Parallel was set to 1. This test would execute until the Parallel() call and be blocked for execution itself then. All other tests that would've been run in the same batch but before this test in that batch would see a non-resolvable domain setting as well.

Using a local variable now to be double sure this doesn't happen again.