[CP] Clean up VirtualService and reconcile kingress when ingress.class annotation is update (#394)
* Clean up VirtualService when ingress.class annotation is not istio (#6589)

When we changed ingress.class annotation, we would like
networking-istio to clean up obsoleted VirtualServices.

This patch changes to skip creating VirtualService which do not have
ingress.class=istio annotation. As a result, obsoleted virtualservices
are cleaned up.

* Reconcile ingress when annotation is updated (#6569)

* Reconcile ingress when annotation is updated

Although `ingress.class` in config-network updates ingress annotation,
reconciler does not work by the difference of annotation.

To fix it, this patch changes to reconcile ingress when annotation is
updated.

* Update unit test