[BACKPORT][0.23] Fix websocket test with https option (#802)
* Do not call GetIngressEndpoint() when --resolvabledomain is enabled (#11318)

`GetIngressEndpoint()` is a function to get ingress endpoint for the
spoof client. So we don't need to call it when `--resolvabledomain` is
enabled. Especially when Ingresss service uses `ClusterIP`, it alwasy
fails to get endpoint and test fails.

This patch changes to call GetIngressEndpoint() only when
resolvabledomain is false in websocket test.

* Use TLS for websocket and gRPC tests (#11322)

* Support websocket https

* Use TLS for websocket and gRPC tests

Current gRPC and Websocket does not use TLS, but still uses HTTP even
when it runs with HTTPS mode.

This patch fixes it.

* Add ServerName to pseudo host with TLS.

* Decrease the length of test name

* Fix race issue

* Add workaround only https

* try dialer option

* Without parallel

* Use a single DialOption

* Revert "Without parallel"

This reverts commit 99455a696fe62d88a6056ccb60993b4d476d83dc.

* Revert "Revert "Without parallel""

This reverts commit 7105cbc99823a24e57ec8a8fb66e30dbd19c00f0.

* Fix TestGRPCStreamingPingViaActivator

* Do not run gRPC with parallel

* Add TODO comment

* Use net.SplitHostPort

* make long ksvc name generic

* Revert "make long ksvc name generic"

This reverts commit 4119f0805f91bdb2366b53c418b506f039cc02a3.