Add k8s minimum version patch (#924)
Upstream does not support Kubernetes 1.19 (OpenShift 4.6) but we still
would like to test our serverless on OCP 4.6.

So this patch adds the patch to `defaultMinimumVersion` in
`vendor/knative.dev/pkg/version/version.go`.

Note, there are other way for the workaround like:
- Add `KUBERNETES_MIN_VERSION` value to each deployment.
- Manage `KUBERNETES_MIN_VERSION` by serverless-operator.

But the patch to `vendor/knative.dev/pkg/version/version.go` is easy
and clear.